prediction loop renaming
