+++ pytorch_lightning/trainer/trainer.py
@@ -28,7 +28,7 @@ from pytorch_lightning.core.lightning import LightningModule
 from pytorch_lightning.core.memory import ModelSummary
 from pytorch_lightning.loggers import LightningLoggerBase
 from pytorch_lightning.loops.dataloader.evaluation_loop import EvaluationLoop
-from pytorch_lightning.loops.dataloader.prediction_dataloader_loop import PredictionDataLoaderLoop
+from pytorch_lightning.loops.dataloader.prediction_loop import PredictionLoop
 from pytorch_lightning.loops.fit_loop import FitLoop
 from pytorch_lightning.plugins import Plugin
 from pytorch_lightning.plugins.environments import ClusterEnvironment
@@ -345,7 +345,7 @@ class Trainer(
         self.fit_loop = FitLoop(min_epochs, max_epochs, min_steps, max_steps)
         self.validation_loop = EvaluationLoop()
         self.test_loop = EvaluationLoop()
-        self.predict_loop = PredictionDataLoaderLoop()
+        self.predict_loop = PredictionLoop()
         self.fit_loop.connect(self)
         self.validation_loop.connect(self)
         self.test_loop.connect(self)
